requirements.txt
streamlit
pydantic
pandas
pytesseract
opencv-python
pdfplumber
matplotlib
seaborn
Pillow
